Fix cfo estimation
The coarse cfo estimation has to be respected more than the pilot based estimation. This is because the pilot based estimation can sometimes loose track of the cfo and produce far off results. Coarse estimation will lead the estimates back to a range where the pilot based cfo can work.